refactor: flatten borderGlow,sessionCard, and collaboratorIncidents widget hierarchies
