+++ packages/client/lib/app/core/widgets/cards/session_card/session_card_store.dart
