+++ packages/client/lib/app/core/widgets/collaborator_presence_incidents_overlay/collaborator_presence_incidents_overlay.dart
@@ -1,1 +1,39 @@
-export 'stack/stack.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_mobx/flutter_mobx.dart';
+import 'package:nokhte/app/core/hooks/hooks.dart';
+import 'package:nokhte/app/core/types/types.dart';
+import 'package:nokhte/app/core/widgets/widgets.dart';
+export 'collaborator_presence_incidents_overlay_store.dart';
+export 'collaborator_presence_incidents_overlay.dart';
+
+class CollaboratorPresenceIncidentsOverlay extends StatelessWidget {
+  final CollaboratorPresenceIncidentsOverlayStore store;
+  const CollaboratorPresenceIncidentsOverlay({
+    super.key,
+    required this.store,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Observer(
+      builder: (context) {
+        return Stack(
+          children: [
+            Padding(
+              padding: const EdgeInsets.only(bottom: 150.0),
+              child: Center(
+                child: AnimatedOpacity(
+                  opacity: useWidgetOpacity(store.showWidget),
+                  duration: Seconds.get(1),
+                  child: const Jost(
+                    "Collaborator Is Offline",
+                  ),
+                ),
+              ),
+            ),
+          ],
+        );
+      },
+    );
+  }
+}
