+++ packages/client/lib/app/core/widgets/collaborator_presence_incidents_overlay/collaborator_presence_incidents_overlay_store.dart
