--- packages/client/lib/app/core/widgets/collaborator_presence_incidents_overlay/stack/presentation/presentation.dart
@@ -1,2 +1,0 @@
-export 'mobx/collaborator_presence_incidents_overlay_store.dart';
-export 'widget/collaborator_presence_incidents_overlay.dart';
