--- packages/client/lib/app/core/widgets/collaborator_presence_incidents_overlay/stack/stack.dart
@@ -1,1 +1,0 @@
-export 'presentation/presentation.dart';
